requirements.txt
numpy==1.22.3
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
mediapipe==0.8.10
mss==6.1.0
Pillow==9.1.1
facenet_pytorch
